+++ scripts/createintegration/main.js
@@ -11,13 +11,13 @@ function main() {
   document.getElementById("checkintegrationexists").addEventListener("click", checkIntegrationPath);
 
   // input fields
-  const fields = document.querySelectorAll(".field,.stepfield,.otherstepfield,.dropdown,.stepdropdown,.existingscenariocheckbox");
+  const fields = document.querySelectorAll(".field,.dropdown,.stepdropdown,.existingscenariocheckbox");
   for (const field of fields) {
     field.addEventListener("keyup", fieldChange);
   }
 
   // fixed/step dropdowns
-  const dropDowns = document.querySelectorAll(".dropdown,.stepdropdown");
+  const dropDowns = document.querySelectorAll(".dropdown,.stepdropdown,.steptypedropdown,.stepmethoddropdown");
   for (const dropDown of dropDowns) {
     dropDown.addEventListener("change", fieldChange);
   }
@@ -26,26 +26,6 @@ function main() {
   const checkBoxes = document.querySelectorAll(".existingscenariocheckbox,.checkallexisting");
   for (const checkbox of checkBoxes) {
     checkbox.addEventListener("change", fieldChange);
-  }
-
-  // on panel creation: save all dropdown values (if exist)
-  saveValue("modulename");
-  saveValue("nofscenarios");
-  if(isCreate()) {
-    saveValue("nofsteps");
-  }
-
-  // stepDropdowns (if create)
-  if(isCreate()) {
-    for (const stepDropDown of document.querySelectorAll(".stepdropdown")) {
-      saveValue(stepDropDown.id);
-  
-      // if 'other': reveal other step field
-      if (stepDropDown.value === 'other') {
-        var index = stepDropDown.getAttribute('indexstep');
-        document.getElementById("otherstepname" + index).style.display = 'inline-table';
-      }
-    }
   }
 
   // scenario grid fields
@@ -70,8 +50,6 @@ function fieldChange(event) {
   switch (field.classList[0]) {
     // input fields: check contents
     case 'field':
-    case 'stepfield':
-    case 'otherstepfield':
       updateFieldOutlineAndTooltip(field.id);
       break;
 
@@ -81,22 +59,6 @@ function fieldChange(event) {
         vscodeApi.postMessage({ command: "clearscenarios", text: '' });
       }
       //vscodeApi.postMessage({ command: "refreshpanel", text: '' });
-      break;
-
-    // stepdropdown: show/hide other step field
-    case 'stepdropdown':
-      var index = field.getAttribute('indexstep');
-      var otherStepField = document.getElementById("otherstepname" + index);
-
-      if (field.value === 'other') {
-        // if 'other': show other step field 
-        otherStepField.style.display = 'inline-table';
-      } else {
-        // not other: hide other step field and delete value
-        otherStepField.style.display = 'none';
-        otherStepField.value = '';
-        saveValue(otherStepField.id);
-      }
       break;
     
     // existingscenariocheckbox: update associated existing scenario field
@@ -127,25 +89,7 @@ function infoMessage(info) {
 
 function saveValue(fieldId) {
   var field = document.getElementById(fieldId);
-  var attr = '';
-  switch (field.classList[0]) {
-    case 'dropdown':
-    case 'field':
-      attr = 'index';
-      break;
-    case 'stepdropdown':
-    case 'stepfield':
-      attr = 'indexstep';
-      break;
-    case 'otherstepfield':
-      attr = 'indexotherstep';
-      break;
-
-    case 'existingscenariocheckbox':
-      attr = 'indexescheckbox';
-      break;
-
-  }
+  var attr = 'index';
   var value = field.checked ?? field.value;
   var textString = field.classList[0] + '|' + (field.getAttribute(attr) ?? '') + '|' + value;
   vscodeApi.postMessage({ command: "savevalue", text: textString });
@@ -154,7 +98,7 @@ function saveValue(fieldId) {
 function checkFields() {
   // check if any incorrect field contents and update fields outlining and tooltip in the process
   var check = true;
-  const fields = document.querySelectorAll(".field,.stepfield,.otherstepfield");
+  const fields = document.querySelectorAll(".field");
   for (const field of fields) {
     check = updateFieldOutlineAndTooltip(field.id) ? check : false;
   }
@@ -183,23 +127,6 @@ function checkContent(id, value) {
           check = 'invalid';
         }
         break;
-      case 'carrierapidescription':
-        check = value.match(/[^A-Za-z0-9\-\:\(\) ]/);
-        break;
-      case 'testuser':
-        check = value.match(/[^A-Za-z0-9\-\_]/);
-        break;
-      case 'testpwd':
-        check = '';
-        break;
-
-      case 'stepfield':
-        check = value.match(/[^A-Za-z0-9\:\/\.\?\=\&\-\_]/);
-        break;
-
-      case 'otherstepfield':
-        check = value.match(/[^a-z0-9\_]/);
-        break;
     }
 
     // if invalid content: return false
@@ -221,23 +148,6 @@ function getContentHint(elementid) {
       break;
     case 'carriercode':
       hint = 'A-Z, 0-9 (only capitals); exactly 3 characters';
-      break;
-    case 'carrierapidescription':
-      hint = 'A-Z, a-z, 0-9, \':\', \'(\', \')\' (spaces allowed)';
-      break;
-    case 'testuser':
-      hint = 'A-Z, a-z, 0-9, \'-\', \'_\' (no spaces)';
-      break;
-    case 'testpwd':
-      hint = 'Any character allowed';
-      break;
-
-    case 'stepfield':
-      hint = 'A-Z, a-z, 0-9, :/.?=&-_ (no spaces)';
-      break;
-
-    case 'otherstepfield':
-      hint = 'a-z, 0-9, \'_\' (no spaces)';
       break;
   }