+++ src/panels/CreateIntegrationHtmlObject.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { getUri, valueString, checkedString, hiddenString, dropdownOptions, arrayFrom1, toBoolean } from "../utilities/functions";
+import { getUri, valueString, checkedString, hiddenString, disabledString, dropdownOptions, arrayFrom1, toBoolean } from "../utilities/functions";
 import { ScenarioGridObject } from "../utilities/ScenarioGridObject";
 
 // fixed fields indices
@@ -7,11 +7,8 @@ const carrierIndex = 0;
 const apiIndex = 1;
 const moduleIndex = 2;
 const carrierCodeIndex = 3;
-const apiDescriptionIndex = 4;
 const nofStepsIndex = 5;
 const nofScenariosIndex = 6;
-const carrierUserIndex = 7;
-const carrierPwdIndex = 8;
 
 export class CreateIntegrationHtmlObject {
   // PROPERTIES
@@ -24,14 +21,19 @@ export class CreateIntegrationHtmlObject {
     private _modularElementsWithParents: {parent:string, element:string, multi:boolean}[],
     private _fieldValues: string[],
     private _stepFieldValues: string[],
-    private _otherStepValues: string[],
     private _scenarioFieldValues: string[],
     private _existingScenarioFieldValues: string[],
     private _existingScenarioCheckboxValues: boolean[],
     private _createUpdateValue: string,
     private _modularValue: boolean,
     private _multiFieldValues: {[details: string] : string;},
-    private _nofPackages: string[]
+    private _nofPackages: string[],
+    private _moduleOptions: string[],
+    private _stepOptions: string[],
+    private _stepTypeOptions: string[],
+    private _stepTypes: string[],
+    private _stepMethodOptions: string[],
+    private _stepMethods: string[]
     ) { }
 
   // METHODS
@@ -83,8 +85,7 @@ export class CreateIntegrationHtmlObject {
             <section class="component-container">
               <div class="component-sub-container">
                 <h2>Carrier</h2>
-                ${this._getCarrierFolderStructureGrid()}
-                ${this._ifCreate(this._getCarrierDetailsGrid())}
+                ${this._getCarrierGrid()}
               </div>
             </section> 
             
@@ -124,15 +125,6 @@ export class CreateIntegrationHtmlObject {
     return outString;
   }
 
-  private _isOther(string: string): boolean {
-    let isOther: boolean = false;
-    if ((string + "").toLowerCase() === 'other') {
-      isOther = true;
-    }
-
-    return isOther;
-  }
-
   private _getCreateUpdateButton(): string {
     let createUpdateButton: string = /*html*/ `
       <vscode-button id="createintegration" appearance="primary" ${this._ifUpdate('style="background-color:green"')}>
@@ -145,7 +137,7 @@ export class CreateIntegrationHtmlObject {
     return createUpdateButton;
   }
 
-  private _getCarrierFolderStructureGrid(): string {
+  private _getCarrierGrid(): string {
     let carrierFolderStructureGrid = /*html*/ `
       <section class="component-example">
         <p>Folder structure:    <b>carrier / api-name / module</b></p>
@@ -154,47 +146,23 @@ export class CreateIntegrationHtmlObject {
         <vscode-text-field id="carrierapiname" class="field" index="${apiIndex}" ${valueString(this._fieldValues[apiIndex])} placeholder="api-name" size="5"></vscode-text-field>
         /
         <vscode-dropdown id="modulename" class="dropdown" index="${moduleIndex}" ${valueString(this._fieldValues[moduleIndex])} position="below">
-          ${dropdownOptions(['booking', 'tracking', 'cancel', 'pickup', 'pickup_cancel'])}
+          ${dropdownOptions(this._moduleOptions)}
         </vscode-dropdown>
 
-        <section class="component-example">
+        <section class="component-nomargin">
           <vscode-button id="checkintegrationexists" appearance="primary">
             Check
             <span slot="start" class="codicon codicon-refresh"></span>
           </vscode-button>
         </section>
+      </section>
+
+      <section class="component-example">
+        <p>SiS CarrierCode</p>
+        <vscode-text-field id="carriercode" class="field" index="${carrierCodeIndex}" ${valueString(this._fieldValues[carrierCodeIndex])} placeholder="DPD"></vscode-text-field>
       </section>`;
 
     return carrierFolderStructureGrid;
-  }
-
-  private _getCarrierDetailsGrid(): string {
-    let carrierDetailsGrid = /*html*/ `
-      <section class="row11">
-        <section class="rowelement">
-          <h4>Carrier details</h4>
-          <section class="component-example">
-            <vscode-text-field id="carriercode" class="field" index="${carrierCodeIndex}" ${valueString(this._fieldValues[carrierCodeIndex])} placeholder="DPD">SiS CarrierCode</vscode-text-field>
-          </section>
-
-          <section class="component-example">
-            <vscode-text-field id="carrierapidescription" class="field" index="${apiDescriptionIndex}" ${valueString(this._fieldValues[apiDescriptionIndex])} placeholder="DPD NL Webservice">Carrier API description</vscode-text-field>
-          </section>
-        </section>
-
-        <section class="rowelement">
-          <h4>Carrier TST credentials</h4>
-          <section class="component-example">
-            <vscode-text-field id="testuser" class="field" index="${carrierUserIndex}" ${valueString(this._fieldValues[carrierUserIndex])} placeholder="DPDTstUser">User</vscode-text-field>
-          </section>
-
-          <section class="component-example">
-            <vscode-text-field id="testpwd" class="field" index="${carrierPwdIndex}" ${valueString(this._fieldValues[carrierPwdIndex])} placeholder="aslfjakl">Pwd</vscode-text-field>
-          </section>
-        </section>
-      </section>`;
-
-    return carrierDetailsGrid;
   }
 
   private _getStepsGrid(): string {
@@ -234,64 +202,54 @@ export class CreateIntegrationHtmlObject {
   }
 
   private _stepInputs(nofSteps: number): string {
-
     let subStepNames: string = '';
-    let subTestUrl: string = '';
-    let subProdUrl: string = '';
+    let subStepTypes: string = '';
+    let subStepMethods: string = '';
     for (let step = 0; step < +nofSteps; step++) {
-      let thisStepFieldValue = this._stepFieldValues[step];
-      let isOther: boolean = this._isOther(thisStepFieldValue);
 
-      // set html string addition
-      let subStepNamesCurrent = /*html*/`
-        <section class="component-example">
-          <vscode-dropdown id="stepname${step}" indexstep="${step}" ${valueString(this._stepFieldValues[step])} class="stepdropdown" position="below">
-            <vscode-option>${(this._fieldValues[moduleIndex] ?? 'booking')}</vscode-option>
-            <vscode-option>label</vscode-option>
-            <vscode-option>login</vscode-option>
-            <vscode-option>get_token</vscode-option>
-            <vscode-option>save_token</vscode-option>
-            <vscode-option>other</vscode-option>
+      // step name dropdown
+      subStepNames += /*html*/`
+        <section class="component-vmargin">
+          <vscode-dropdown id="stepname${step}" index="${step}" ${valueString(this._stepFieldValues[step])} class="stepdropdown" position="below">
+            <vscode-option>${this._fieldValues[moduleIndex]}</vscode-option>
+            ${dropdownOptions(this._stepOptions)}
           </vscode-dropdown>
-
-          <vscode-text-field id="otherstepname${step}" indexotherstep="${step}" ${valueString(this._otherStepValues[step])} class="otherstepfield" placeholder="step" ${hiddenString(isOther)}></vscode-text-field>
         </section>
       `;
 
-      // replace nth <vscode-option> occurrence to pre-set different selected for each step 
-      // BUT: only if not already set in _stepFieldValues
-      if (this._stepFieldValues[step] === undefined) {
-        // from https://stackoverflow.com/a/44568739/1716283
-        let t: number = 0;
-        subStepNamesCurrent = subStepNamesCurrent.replace(/<vscode-option>/g, match => ++t === (step + 1) ? '<vscode-option selected>' : match);
-      }
 
-      subStepNames += subStepNamesCurrent;
+      // step type dropdown
+      subStepTypes += /*html*/`
+        <section class="component-vmargin">
+          <vscode-dropdown id="steptype${step}" index="${step}" ${valueString(this._stepTypes[step])} class="steptypedropdown" position="below">
+            ${dropdownOptions(this._stepTypeOptions)}
+          </vscode-dropdown>
+        </section>
+      `;
 
-      subTestUrl += /*html*/ `
-        <section class="component-example">
-          <vscode-text-field id="testurl${step}" indexstep="${step + 10}" ${valueString(this._stepFieldValues[step + 10])} class="stepfield" placeholder="https://test-dpd.com/booking"></vscode-text-field>
-        </section>`;
-
-      subProdUrl += /*html*/ `
-        <section class="component-example">
-          <vscode-text-field id="produrl${step}" indexstep="${step + 20}" ${valueString(this._stepFieldValues[step + 20])} class="stepfield" placeholder="https://prod-dpd.com/booking"></vscode-text-field>
-        </section>`;
+      // step method dropdown
+      subStepMethods += /*html*/`
+        <section class="component-vmargin">
+          <vscode-dropdown id="stepmethod${step}" index="${step}" ${valueString(this._stepMethods[step])} class="stepmethoddropdown" ${disabledString(this._stepTypes[step] === 'http')} position="below">
+            ${dropdownOptions(this._stepMethodOptions)}
+          </vscode-dropdown>
+        </section>
+      `;
     }
 
     let html: string = /*html*/ `
       <section class="component-example">
         <section class="floatleftnopadding">
-          <p>step name</p>
+          <p>Name</p>
           ${subStepNames}
         </section>
         <section class="floatleftnopadding">
-          <p>test url</p>
-          ${subTestUrl}
+          <p>Type</p>
+          ${subStepTypes}
         </section>
         <section class="floatleftnopadding">
-          <p>prod url</p>
-          ${subProdUrl}
+          <p>Method</p>
+          ${subStepMethods}
         </section>
       </section>`;
 
@@ -311,7 +269,7 @@ export class CreateIntegrationHtmlObject {
 
       html += /*html*/`
         <section class="component-example">
-          <vscode-checkbox id="runexistingscenario${index}" class="existingscenariocheckbox" indexescheckbox="${index}" ${checked}></vscode-checkbox>
+          <vscode-checkbox id="runexistingscenario${index}" class="existingscenariocheckbox" index="${index}" ${checked}></vscode-checkbox>
           <vscode-text-field id="existingscenario${index}" class="existingscenariofield" value="${this._existingScenarioFieldValues[index]}" ${disabledReadonly}></vscode-text-field>
         </section>
       `;